+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Define borehole csv import fields `original_name`, `location_x` and `location_y` as required.
 - Add validation for required borehole csv import fields. 
 - Added online help.
+- Add validation for duplicated boreholes during the cvs import.
 
 ### Changed